Added hooks class to project
